Move `generic_type_registry` under `Tapioca::Runtime`
