+++ lib/tapioca/internal.rb
@@ -5,7 +5,7 @@ require "tapioca"
 require "tapioca/loader"
 require "tapioca/sorbet_ext/generic_name_patch"
 require "tapioca/sorbet_ext/fixed_hash_patch"
-require "tapioca/generic_type_registry"
+require "tapioca/runtime/generic_type_registry"
 require "tapioca/helpers/cli_helper"
 require "tapioca/helpers/config_helper"
 require "tapioca/helpers/rbi_helper"
